requirements.txt
tensorflow>=2.10.0
streamlit>=1.28.0
numpy>=1.23.0
pillow>=9.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
opencv-python>=4.7.0
pandas>=1.5.0
plotly>=5.13.0
tqdm>=4.65.0
